Make gemini workable
