+++ internal/integrations/gemini/schema.go
@@ -1,17 +1,13 @@
 package gemini
 
 import (
-	"context"
 	"fmt"
-	"strings"
 
-	"github.com/vlatan/video-store/internal/drivers/rdb"
-	"github.com/vlatan/video-store/internal/models"
 	"google.golang.org/genai"
 )
 
 // produceSchema defines the JSON schema for the response
-func (s *Service) responseSchema(ctx context.Context) *genai.Schema {
+func (s *Service) responseSchema() *genai.Schema {
 	return &genai.Schema{
 		Type: genai.TypeObject,
 		Properties: map[string]*genai.Schema{
@@ -28,33 +24,10 @@ func (s *Service) responseSchema(ctx context.Context) *genai.Schema {
 				Type: genai.TypeString,
 				Description: fmt.Sprintf(
 					"Select only ONE category from these categories: %s.",
-					s.catString(ctx),
+					s.catStr,
 				),
 			},
 		},
 		Required: []string{"title", "summary", "category"},
 	}
-}
-
-// catString creates a string of categories separated by comma
-func (s *Service) catString(ctx context.Context) string {
-
-	// Get the categories from cache or DB
-	categories, _ := rdb.GetCachedData(
-		ctx,
-		s.rdb,
-		"categories",
-		s.config.CacheTimeout,
-		func() (models.Categories, error) {
-			return s.catsRepo.GetCategories(ctx)
-		},
-	)
-
-	// Extract the category names
-	catNames := make([]string, len(categories))
-	for i, cat := range categories {
-		catNames[i] = cat.Name
-	}
-
-	return strings.Join(catNames, ", ")
 }
